feat : RowMapperImpl @FunctionalInterface 적용
